Clases/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
openai==1.12.0
PyPDF2==3.0.1
